+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 en-US:
 
 - Add setting main window theme color feature
+- Add status bar display session information feature
 
 zh-CN:
 
 - 增加设置主界面主题色功能
+- 增加状态栏显示会话信息功能
 
 ## [[V0.4.5](https://github.com/QQxiaoming/quardCRT/releases/tag/V0.4.5)] - 2024-04-26